// brick-erp-basic-setup/types.ts
export type GlobalSettingsData = {
  id: string;
  salesPricePerUnit: number;
  maintenanceReservePerUnit: number;
  laborRate: number;
  cementUnitCost: number;
  sandUnitCost: number;
  dieselUnitCost: number;
  cementConversionFactor: number;
  sandConversionFactor: number;
  dieselConversionFactor: number;
  cementRatio: number;
  sandRatio: number;
  dieselRatio: number;
  updatedById: string;
  updatedAt: Date;
};